+++ src/presets/builtin/index.ts
@@ -1,2 +1,3 @@
 export {DefaultEditorPreset} from './DefaultEditorPreset';
 export {SafeMarkdownPreset} from './SafeMarkdownPreset';
+export {StarterKitPreset} from './StarterKitPreset';