+++ vitest.config.ts
@@ -1,10 +1,12 @@
-import {defineConfig} from 'vitest/config';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
     test: {
         environment: 'jsdom',
         globals: true,
-        include: ['tests/**/*.test.ts'],
+        include: [
+            'tests/**/*.{test,spec}.ts'
+        ],
         setupFiles: ['./tests/setup.ts'],
         coverage: {
             reporter: ['text', 'html']